\name{RcppInt64-package}
\alias{RcppInt64-package}
\alias{RcppInt64}
\docType{package}
\title{\packageTitle{RcppInt64}}
\description{\packageDescription{RcppInt64}}
\section{Package Content}{\packageIndices{RcppInt64}}
\author{\packageAuthor{RcppInt64}}
\section{Maintainer}{\packageMaintainer{RcppInt64}}
\keyword{package}
